% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{compute_bottom_up_fluxes}
\alias{compute_bottom_up_fluxes}
\title{Compute bottom-up fluxes}
\usage{
compute_bottom_up_fluxes(M, nIt = 100)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}

\item{nIt}{Number of iterations for the simulation.}
}
\value{
A matrix of accumulated fluxes.
}
\description{
Simulates bottom-up biomass fluxes through a trophic network
using an iterative procedure.
}
