% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_connectance}
\alias{compute_connectance}
\title{Connectance (C)}
\usage{
compute_connectance(M, exclude_self = FALSE)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}

\item{exclude_self}{Logical, whether to exclude self-links.}
}
\value{
Numeric, connectance.
}
\description{
Fraction of realized links out of all possible links.
}
