% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_inward_degree}
\alias{compute_inward_degree}
\title{Compute inward degree}
\usage{
compute_inward_degree(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
A numeric vector of inward degrees.
}
\description{
Computes the inward degree of each node as the number of incoming links.
}
