% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_L}
\alias{compute_L}
\title{Number of links (L)}
\usage{
compute_L(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
Integer, number of links.
}
\description{
Total number of feeding interactions in the food web.
}
