% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_linkage_density}
\alias{compute_linkage_density}
\title{Linkage density (L/S)}
\usage{
compute_linkage_density(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
Numeric, linkage density.
}
\description{
Average number of links per nodes.
}
