% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_outward_degree}
\alias{compute_outward_degree}
\title{Compute outward degree}
\usage{
compute_outward_degree(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
A numeric vector of outward degrees.
}
\description{
Computes the outward degree of each node as the number of outgoing links.
}
