% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{compute_S}
\alias{compute_S}
\title{Number of nodes (S)}
\usage{
compute_S(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
Integer, number of nodes.
}
\description{
Number of nodes in the food web.
}
