% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.r
\name{get_basal_nodes}
\alias{get_basal_nodes}
\title{Identify basal nodes}
\usage{
get_basal_nodes(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
An integer vector of indices corresponding to basal nodes.
}
\description{
Returns the indices of basal nodes in an interaction matrix.
Basal nodes are those with no incoming links (column sum equal to zero).
}
