% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{get_leaf_nodes}
\alias{get_leaf_nodes}
\title{Identify leaf nodes}
\usage{
get_leaf_nodes(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
An integer vector of indices corresponding to leaf nodes.
}
\description{
Returns the indices of leaf nodes in an interaction matrix.
Leaf nodes are those with no outgoing links (row sum equal to zero).
}
