% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_builder.R
\name{unflatten_foodweb}
\alias{unflatten_foodweb}
\title{Reconstruct a food web adjacency matrix from long format}
\usage{
unflatten_foodweb(foodweb_tab)
}
\arguments{
\item{foodweb_tab}{A data frame in long format with at least three columns:
\code{prey}, \code{consumer}, and \code{interaction}.}
}
\value{
A matrix with prey as row names, consumers as column names, and
\code{interaction} values as the matrix entries.
}
\description{
Converts a flattened food web data frame back into an adjacency matrix.
}
\details{
This function assumes that each prey–consumer pair appears at most once in
\code{foodweb_tab}.
}
\examples{
\dontrun{
## Fake foodweb table
prey <- rep(c("a","b","c"), each=3)
consumers <- rep(c("A","B","C"))
interactions <- rnorm(9)
foodweb_tab <- data.frame(prey=prey, consumer=consumers, interaction=interactions)

## Visualise
print(foodweb_tab)

## Unflatten table
foodweb_mat <- unflatten_foodweb(foodweb_tab)

## Visualise
print(foodweb_mat)
}

}
\seealso{
\code{\link[=flatten_foodweb]{flatten_foodweb()}}
}
