% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networks.R
\name{compute_inward_degree}
\alias{compute_inward_degree}
\title{Compute inward degree}
\usage{
compute_inward_degree(M)
}
\arguments{
\item{M}{A square adjacency or interaction matrix.}
}
\value{
A numeric vector of inward degrees.
}
\description{
Computes the inward degree (column sums) of each node.
}
